Launch cassandra with exec so it can respond to SIGTERM
The entrypoint for our containers is a bash startup script. When you stop
the container, SIGTERM is sent to the script and not the cassandra process.
Cassandra would shut down gracefully if it received the signal, but it does not,
so we have to wait for the operation to timeout and then docker will kill the
container. This change launches the containers with 'exec' so that the
cassandra process becomes pid 1, and therefore receives the signals, and shuts
down immediately.